Rename: [EVER-93] 폴더명 수정 types.ts -> types
